pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aind-ibl-ephys-alignment-preprocessing"
description = "Library to prepare histology and ephys for the IBL ephys alignment GUI"
license = "MIT"
requires-python = ">=3.10"
authors = [
  { name = "Allen Institute for Neural Dynamics" },
  { name = "Galen Lynch", email = "galen@galenlynch.com" },
]
classifiers = ["Programming Language :: Python :: 3"]
readme = "README.md"
version = "0.1.0"

dependencies = [
  "numpy>=1.21.0",
  "pandas>=1.3.0",
  "pydantic>=2",
  "antspyx>=0.3.0",
  "iblatlas>=0.4.0",
  "SimpleITK",
  "aind-anatomical-utils>=0.4,<0.5",
  "aind-ephys-ibl-gui-conversion @ git+https://github.com/AllenNeuralDynamics/aind-ephys-ibl-gui-conversion@main",
  "aind-registration-utils>=0.4,<0.5",
  "aind-s3-cache>=0.1",
  "aind-zarr-utils>=0.11.2",
  "filelock",
]

[project.urls]
Repository = "https://github.com/AllenNeuralDynamics/aind-ibl-ephys-alignment-preprocessing/"
Issues = "https://github.com/AllenNeuralDynamics/aind-ibl-ephys-alignment-preprocessing/issues"
Changelog = "https://github.com/AllenNeuralDynamics/aind-ibl-ephys-alignment-preprocessing/blob/main/CHANGELOG.md"

[project.scripts]
aind-ibl-preprocess = "aind_ibl_ephys_alignment_preprocessing.scripts.run:main"

[dependency-groups]
types = ["mypy"]
dev = [{ include-group = "types" }, "codespell", "pytest", "pytest-cov", "ruff", "interrogate"]
jupyter = [{ include-group = "dev" }, "ipykernel", "pip"]
docs = [
  { include-group = "dev" },
  "sphinx>=5.0",
  "sphinx-autodoc-typehints",
  "sphinx-copybutton",
  "myst-parser",
  "furo",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/aind_ibl_ephys_alignment_preprocessing"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "I", "F", "E", "W", "D", "UP", "ANN", "PYI"]
extend-ignore = ["ANN401"]  # Skip self/cls type annotations
# Skip all type annotation checks in tests
per-file-ignores = { "tests/*.py" = ["ANN"] }
mccabe = { max-complexity = 14 }
pydocstyle = { convention = "numpy" }

[tool.codespell]
skip = ".git,*.pdf,*.svg,./docs/_build,uv.lock"
ignore-words-list = ""

[tool.pytest.ini_options]
addopts = "--strict-markers --tb=short --cov=src --cov-report=term-missing --cov-fail-under=0"
testpaths = ["tests"]

[tool.interrogate]
ignore-init-method = true
ignore-magic = true
ignore_module = true
fail-under = 30
exclude = [
  "__init__.py",
  "tests",
  "docs",
  "build",
  "examples",
  "site",
]

[tool.commitizen]
version_provider = "pep621"
update_changelog_on_bump = true
tag_format = "v$version"
major_version_zero = true

[tool.mypy]
files = ["src/aind_ibl_ephys_alignment_preprocessing"]
strict = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = [
  "ants",
  "ants.*",
  "SimpleITK",
  "SimpleITK.*",
  "aind_anatomical_utils.*",
  "aind_ephys_ibl_gui_conversion.*",
  "aind_registration_utils.*",
  "aind_s3_cache.*",
  "aind_zarr_utils.*",
  "iblatlas.*",
  "filelock",
]
ignore_missing_imports = true

